Move post component to posts directory
